+++ windows/UCHandler.h
@@ -72,7 +72,7 @@ public:
 				subMenu.CreatePopupMenu();				
 				subMenu.InsertSeparatorLast(TSTRING(SETTINGS_USER_COMMANDS));
 				
-				menu.AppendMenu(MF_POPUP, (UINT)(HMENU)subMenu, CTSTRING(SETTINGS_USER_COMMANDS));
+				menu.AppendMenu(MF_POPUP, (HMENU)subMenu, CTSTRING(SETTINGS_USER_COMMANDS));
 			}
 			
 			CMenuHandle cur = SETTING(UC_SUBMENU) ? subMenu.m_hMenu : menu.m_hMenu;
